Use nipy_spectral colormap
cm.spectral was renamed a while ago to cm.nipy_spectral with
a deprecation warning, and the original name was later removed. This
makes the writeData script work with current matplotlib versions. It
should also be compatible with older releases, as the nipy_spectral
alias was created in version 1.3, released in 2013.